starter_code/backend/requirements.txt
astroid
python-jose
Click
ecdsa
Flask
Flask-SQLAlchemy
future
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
pylint
six
typed-ast
Werkzeug
wrapt
Flask-Cors